# [feature](pipeline) avoid unnecessary copying during projection.