--- a/be/src/vec/columns/column_map.cpp
+++ b/be/src/vec/columns/column_map.cpp
@@ -58,10 +58,12 @@ ColumnMap::ColumnMap(MutableColumnPtr&& keys, MutableColumnPtr&& values, Mutable
 
         /// This will also prevent possible overflow in offset.
         if (keys_column->size() != last_offset) {
-            LOG(FATAL) << "offsets_column has data inconsistent with key_column";
+            throw Exception(ErrorCode::INTERNAL_ERROR,
+                            "offsets_column has data inconsistent with key_column");
         }
         if (values_column->size() != last_offset) {
-            LOG(FATAL) << "offsets_column has data inconsistent with value_column";
+            throw Exception(ErrorCode::INTERNAL_ERROR,
+                            "offsets_column has data inconsistent with key_column");
         }
     }
 }